requirements.txt
django==1.11.14
mysqlclient
pillow
